requirements.txt
pandas
numpy
scikit-learn
SQLAlchemy
dash
dash-bootstrap-components
mysql-connector-python
dash-core-components
dash-html-components
folium